requirements.txt
numpy
pandas
scikit-learn
streamlit  
requests
groq
python-dotenv
